Adds flag datacenterId as byte into blobid.
Test done:
./gradlew clean && ./gradlew build

Test coverage:
Class Class, %  Method, % Line, %
BlobId  100% (1/ 1) 88.9% (16/ 18)  81.5% (88/ 108)
BlobIdBuilder 100% (1/ 1) 100% (8/ 8) 100% (19/ 19)

+++ ambry-clustermap/src/main/java/com.github.ambry.clustermap/Datacenter.java
@@ -33,7 +33,7 @@ import org.slf4j.LoggerFactory;
  */
 public class Datacenter {
   // datacenterId = LEGACY_DATACENTER_ID indicate datacenterId is not available at the time when this blobId is formed.
-  public static final short LEGACY_DATACENTER_ID = -1;
+  public static final byte LEGACY_DATACENTER_ID = -1;
   private final HardwareLayout hardwareLayout;
   private final String name;
   private final ArrayList<DataNode> dataNodes;
